Move Bank-based tests into unit-tests
